apps/backend/requirements.txt
accelerate==1.3.0
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
asyncpg==0.30.0
attrs==25.1.0
beautifulsoup4==4.13.0
bioc==2.1
boto3==1.36.11
botocore==1.36.11
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
conllu==4.5.3
contourpy==1.3.1
cycler==0.12.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
filelock==3.17.0
flair==0.15.0
fonttools==4.55.8
fsspec==2025.2.0
ftfy==6.3.1
gdown==5.2.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
iniconfig==2.0.0
intervaltree==3.1.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
kiwisolver==1.4.8
langdetect==1.0.9
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
more-itertools==10.6.0
mpld3==0.5.10
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pptree==3.1
protobuf==5.29.3
psutil==6.1.1
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
pytest==8.3.4
pytest-asyncio==0.25.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytorch_revgrad==0.2.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
s3transfer==0.11.2
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
segtok==1.5.11
sentencepiece==0.2.0
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlitedict==2.1.0
starlette==0.45.3
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformer-smaller-training-vocab==0.4.0
transformers==4.48.2
triton==3.2.0
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==14.2
Wikipedia-API==0.8.1
wrapt==1.17.2
